infinite depth
more conditions on using the cache of finished states
seen set of states keeps track of the states seen, and the depth they were seen, when a duplicate state is seen, it calls a minimax search of depth check_tie_depth, which is 5 according to wikipedia.
keeps track of known tie states for fast evaluation. not sure if this is redundant with the finished cache
pruning gives wrong answers (10/100k times) but it is faster on the first turn. however, since it skips evaluating a lot of states, the experiments are actually slower.